Refactored date utils out of timewarp.py
